services/stream-processor/requirements.txt
# Kafka client
kafka-python==2.0.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Caching
redis==5.0.1

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
python-json-logger==2.0.7